+++ newton/examples/cloth/example_cloth_rollers.py
@@ -362,11 +362,15 @@ class Example:
             model=self.model,
             iterations=self.iterations,
             particle_enable_self_contact=True,
-            particle_self_contact_radius=0.3,
-            particle_self_contact_margin=0.6,
+            particle_self_contact_margin=0.3,
+            particle_self_contact_gap=0.3,
             particle_vertex_contact_buffer_size=48,
             particle_edge_contact_buffer_size=64,
-            particle_collision_detection_interval=5,
+            collision_frequency=[1, 5],
+            collision_frequency_type=[
+                newton.solvers.SolverBase.CollisionFrequencyType.AUTO,
+                newton.solvers.SolverBase.CollisionFrequencyType.ITERATIONS,
+            ],
             particle_topological_contact_filter_threshold=2,
         )